refactor: move custom compilers to _internal
